Decompose schema-related source files
