+++ src/charge/uc-value-builder.ts
@@ -1,6 +1,6 @@
-import { UcEntity } from '../schema/uc-entity.js';
-import { UcList } from '../schema/uc-list.js';
-import { UcMap } from '../schema/uc-map.js';
+import { UcEntity } from '../schema/entity/uc-entity.js';
+import { UcList } from '../schema/list/uc-list.js';
+import { UcMap } from '../schema/map/uc-map.js';
 import { UcPrimitive } from '../schema/uc-primitive.js';
 import { UcValue } from '../schema/uc-value.js';
 import { OpaqueURIChargeRx } from './opaque.uri-charge-rx.js';
