+++ src/compiler/serialization/default.ucs-defs.ts
@@ -2,8 +2,8 @@ import { noop } from '@proc7ts/primitives';
 import { encodeUcsKey } from '../../impl/encode-ucs-string.js';
 import { SERIALIZER_MODULE } from '../../impl/module-names.js';
 import { escapeJsString } from '../../impl/quote-property-key.js';
-import { UcList } from '../../schema/uc-list.js';
-import { UcMap } from '../../schema/uc-map.js';
+import { UcList } from '../../schema/list/uc-list.js';
+import { UcMap } from '../../schema/map/uc-map.js';
 import { ucNullable } from '../../schema/uc-nullable.js';
 import { ucOptional } from '../../schema/uc-optional.js';
 import { ucSchemaName } from '../../schema/uc-schema-name.js';
