+++ src/readme.spec.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from '@jest/globals';
 import { chargeURI } from './charge/charge-uri.js';
 import { churi } from './charge/churi.tag.js';
-import { UcEntity } from './schema/uc-entity.js';
+import { UcEntity } from './schema/entity/uc-entity.js';
 
 const DOC_EXAMPLE =
   'https://example.com/api(!v:3.0)/user;id=0n302875106592253'
