+++ src/schema/entity/mod.ts
@@ -1,0 +1,1 @@
+export * from './uc-entity.js';
