+++ src/schema/map/mod.ts
@@ -1,0 +1,1 @@
+export * from './uc-map.js';
