+++ src/schema/map/uc-map.ts
@@ -1,9 +1,9 @@
 import { asis } from '@proc7ts/primitives';
-import { jsPropertyKey } from '../impl/quote-property-key.js';
-import { UcPrimitive } from './uc-primitive.js';
-import { ucSchemaName } from './uc-schema-name.js';
-import { UcSchema, UcSchema__symbol } from './uc-schema.js';
-import { UcValue } from './uc-value.js';
+import { jsPropertyKey } from '../../impl/quote-property-key.js';
+import { UcPrimitive } from '../uc-primitive.js';
+import { ucSchemaName } from '../uc-schema-name.js';
+import { UcSchema, UcSchema__symbol } from '../uc-schema.js';
+import { UcValue } from '../uc-value.js';
 
 /**
  * URI charge map represented as JavaScript object.
