+++ src/schema/mod.ts
@@ -1,8 +1,11 @@
+export * from './boolean/mod.js';
+export * from './entity/mod.js';
+export * from './list/mod.js';
+export * from './map/mod.js';
+export * from './numeric/mod.js';
+export * from './string/mod.js';
 export * from './uc-deserializer.js';
-export * from './uc-entity.js';
 export * from './uc-error.js';
-export * from './uc-list.js';
-export * from './uc-map.js';
 export * from './uc-nullable.js';
 export * from './uc-optional.js';
 export * from './uc-primitive.js';
@@ -10,5 +13,5 @@ export * from './uc-schema-name.js';
 export * from './uc-schema-resolver.js';
 export * from './uc-schema.js';
 export * from './uc-serializer.js';
-export * from './uc-unknown.js';
 export * from './uc-value.js';
+export * from './unknown/mod.js';
