+++ src/schema/numeric/uc-bigint.ts
@@ -1,0 +1,9 @@
+import { UcSchema } from '../uc-schema.js';
+
+export type UcBigInt = bigint;
+
+export namespace UcBigInt {
+  export interface Schema extends UcSchema<bigint> {
+    readonly type: typeof BigInt;
+  }
+}
