+++ src/schema/string/mod.ts
@@ -1,0 +1,1 @@
+export * from './uc-string.js';
